# packages/openapi/v1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "description": "Creates a new user account.",
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserHttpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "409": {
            "description": "User with this email already exists"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "description": "Returns paginated list of users.",
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Returns detailed user information.",
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      },
      "put": {
        "description": "Updates user profile information.",
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserHttpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "description": "Marks user as deleted (can be restored).",
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User soft deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Soft delete user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/restore": {
      "patch": {
        "description": "Restores a previously soft-deleted user.",
        "operationId": "UserController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User restored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Restore soft-deleted user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/hard": {
      "delete": {
        "description": "⚠️ IRREVERSIBLE - Completely removes user from database.",
        "operationId": "UserController_hardDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User permanently deleted"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Permanently delete user",
        "tags": [
          "users"
        ]
      }
    },
...
    "/admin/users/{id}": {
      "patch": {
        "operationId": "AdminController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "ویرایش کاربر (status، isAdmin، role)",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "حذف نرم کاربر",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Xennic Platform API",
    "description": "Xennic Engineering Platform API Documentation",
    "version": "1.0.0",
    "contact": {
      "name": "Xennic Team",
      "url": "https://xennic.com",
      "email": "support@xennic.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://xennic.com/terms"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateUserHttpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name",
            "minLength": 2,
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name",
            "minLength": 2,
            "maxLength": 50
          },
          "password": {
            "type": "string",
            "example": "SecurePass123!",
            "description": "Password"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890",
            "description": "Phone number"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "status",
          "isEmailVerified",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserHttpDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890",
            "description": "Phone number"
          }
        }
      },
...
    }
  }
}